+++ extensions-leaflet-map.php
@@ -5,7 +5,7 @@
  * GitHub Plugin URI: https://github.com/hupe13/extensions-leaflet-map-github
  * Primary Branch:    main
  * Description:       Extensions for the WordPress plugin Leaflet Map Github Version
- * Version:           3.4.3-230422
+ * Version:           3.4.4-230426
  * Requires PHP:      7.4
  * Author:      hupe13
  * License:     GPL v2 or later
@@ -65,8 +65,8 @@ include_once LEAFEXT_PLUGIN_DIR . '/php/choropleth.php';
 
 include_once LEAFEXT_PLUGIN_DIR . '/php/zoomhome.php';
 //include_once LEAFEXT_PLUGIN_DIR . '/php/zoomhome-a.php';
-include_once LEAFEXT_PLUGIN_DIR . '/php/layerswitch.php';
-include_once LEAFEXT_PLUGIN_DIR . '/php/providers.php';
+
+include_once LEAFEXT_PLUGIN_DIR . '/php/tileserver.php';
 
 include_once LEAFEXT_PLUGIN_DIR . '/php/leaflet-search.php';