+++ leaflet-plugins/leaflet-elevation-2.3.4/src/options.js
@@ -14,6 +14,7 @@ export var Options = {
 	dragging: !L.Browser.mobile,
 	downloadLink: 'link',
 	elevationDiv: "#elevation-div",
+	edgeScale: { bar: true, icon: false, coords: false },
 	followMarker: true,
 	imperial: false,
 	legend: true,
@@ -81,4 +82,7 @@ export var Options = {
 	interpolation: "curveLinear",
 	yAxisMax: undefined,
 	yAxisMin: undefined,
+
+	// Prevent CORS issues for relative locations (dynamic import)
+	srcFolder: ((document.currentScript && document.currentScript.src) || (import.meta && import.meta.url)).split("/").slice(0,-1).join("/") + '/',
 };