+++ changes.md
@@ -1,4 +1,6 @@
-* none
+* leaflet-elevation-2.3.4
+* Raruto fixed [Map automatically moves after clicking on markers when rotate: true](https://github.com/Raruto/leaflet-elevation/issues/250)
+* layerswitch: providers can have a [mapid](https://wordpress.org/support/topic/change-maps-name-in-layerswitch/)
 
 Please note the [install instructions](https://leafext.de/en/doku/about/versions/).